chore: Relocate GQL schema codes related to Image (#2938)
Backported-from: main (24.12)
Backported-to: 24.09
Backport-of: 2938
